Refactor how inference tasks are run within docker
Refactors the organization and error messages for running
inference-related tasks from Taskfile. Tasks should now be 
run in docker locally using `task docker-run`.

This is because when the tasks are running in docker on CI,
they should not be calling out to the `docker` command itself:
the environment is already within docker.