Add Inference Tasks to CI (#871)
* Export IS_DOCKER from taskcluster/docker/base/Dockerfile

Ensures that the `IS_DOCKER` environment variable is set to `1`
for all CI tasks, by exporting it within the base Dockerfile build.

* Define Dockerfile for inference-related tasks in CI

Defines a Dockerfile for an image that is capable of running
tasks related to the inference engine in the taskcluster CI.

* Refactor how inference tasks are run within docker

Refactors the organization and error messages for running
inference-related tasks from Taskfile. Tasks should now be 
run in docker locally using `task docker-run`.

This is because when the tasks are running in docker on CI,
they should not be calling out to the `docker` command itself:
the environment is already within docker.

* Add taskcluster kind for inference

Adds a taskcluster kind definition for tasks to run
related to the inference directroy on each PR.

Ensures that these tasks only run if releavant files
to these tasks have been modified.

* Rename "all*" kinds to "all*-pipeline"

Renames taskcluster kinds that use the wording "all" to be 
called "all-pipeline" to improve the clarity of what they do.