Remove unused WASM test-page files
This removes the files for building a WASM test page that
functions very similarly to the `about:translations` page
in Firefox.

I kept these files around during the initial clone of the
repository, because I wasn't sure yet if I wanted to use
anything here for testing in CI, but I think it will be
cleaner to just make new CI tests.