Setup WASM test infrastructure for CI (#920)
* Rename task `inference-test` to `inference-test-local`

This commit clarifies that these tests are the local C++
unit tests, now that we will be adding WASM JS tests to CI.

* Remove unused WASM test-page files

This removes the files for building a WASM test page that
functions very similarly to the `about:translations` page
in Firefox.

I kept these files around during the initial clone of the
repository, because I wasn't sure yet if I wanted to use
anything here for testing in CI, but I think it will be
cleaner to just make new CI tests.

* Add vitest infrastructure for WASM CI tests

https://vitest.dev/ seems like a simple, easy-to-use
JS testing famework that is installable and configurable
with NPM.

This patch only introduces a stubbed test file with a basic
assertion, but I plan to use this to test the WASM bindings
and outputs more thoroughly in a subsequent PR.

* Add task `inference-test-wasm`

This commit adds a new task for `inference-test-wasm` that also
runs in CI on relevant PRs that touch inference-related sections
of the code.

* Ensure macOS host dependencies build on 1 thread

There is an issue with building WASM using multiple
threads on Aarch64 macOS chips. It works on a single
thread. This patch exposes the host operating system
type to the Docker container so that the `inference-test-wasm`
task can determine how many threads to use.