Add Web Worker simulation infrastructure
This commit adds a minimal API surface of the WorkerGlobalScope
API functionality that we use for Translations within Firefox,
wrapping the Node.js worker_threads equivalent behavior underneath.

This allows us to test the generated code in a Node.js environment
with the same API calls that we use in Firefox.