Rename bergamot-translator directive to bergamot-translator-source
The name of the file that we use in the mozilla-unified source tree
is `bergamot-translator.js`, but the name of the file generated here
is `bergamot-translator-worker.js`.

I wanted the names to match, so I am renaming the CMake directives
that dictate the generated file's names such that the generated
WASM code will be `bergamot-translator.js`.

This is the first step of that process.