Test WASM Translations in CI (#927)
* Consolidate .gitignore file to inference directory

This commit consolidates all preexisting inference-related gitignore
directives into the `.gitignore` file in the `inference` directory.

* Add eslint to project

This commit adds an eslint config for the WASM tests.
I think this overall adheres to Mozilla's code-format
preferences and is good enough for a first pass.

I have found the linting config to be a bit finnicky,
so my preference would be to improve the linting in a
follow-up, if needed/desired at all.

* Add new lint tasks for eslint

This commit adds new tasks to run the eslint linter
on relevant JavaScript files in the project, as well
as hooks up the tasks to a kind.yml file to run in CI.

* Rename bergamot-translator directive to bergamot-translator-source

The name of the file that we use in the mozilla-unified source tree
is `bergamot-translator.js`, but the name of the file generated here
is `bergamot-translator-worker.js`.

I wanted the names to match, so I am renaming the CMake directives
that dictate the generated file's names such that the generated
WASM code will be `bergamot-translator.js`.

This is the first step of that process.

* Rename bergamot-translator-worker directive to bergamot-tarnslator

This is the second step of the previous commit, which renames
the WASM-related directives to simply be `bergamot-translator`.

This results in the generated JavaScript file being `bergamot-translator.js`
instead of `bergamot-translator-worker.js`.

* Move thread-count default logic into build_wasm.py

Given the issue where building the WASM within the Docker container 
fails on multiple threads only if the host operating system is macOS,
I have moved that default logic within the script itself. 

The default can still be overridden by passing the `-j` flag, but rather
than call sites having to know to do the "right" thing for macOS, I'm making 
it the default intrinsic behavior within the script.

* Prepare Bergamot module for mozilla-unified in build-bergamot.py

This moves the logic that is currently in the mozilla-unified tree,
of adding the licensing, and wrapping the generated WASM JavaScript
module in a function.

This will be paired with a downstream-pr that removes this step on
the mozilla-unified end.

* Add Typescript bindings the for Bergamot

This commit adds some Typescript bindings to the
test directory that match the generated JS.

I spent some time trying to get emscripten to generate
these automatically, but I gave up on my time-boxed effort.

* Add support for `git-lfs` to base docker image

This commit adds support for pulling files via `git-lfs`
to the Dockerfile for the base docker image. In order
to pull the files, we need to install `git-lfs` from apt,
but also add github.com to the list of known ssh hosts.

* Add a subset of models for testing using `git-lfs`

This commit adds the gzipped artifacts for

* `enes`
* `enfr`
* `esen`
* `fren`

These are used for testing for the moment, but I view this as a
temporary solution that is good enough for this PR.

In the future, we will need to merge the `firefox-translations-models`
repository here.

* Add test-wasm.py script

This commit adds a Python script for testing the WASM,
which runs the WASM build script (if needed), and then
invokes the test runner.

* Extract test models from archives in test-wasm.py

This commit modifies the new `test-wasm.py` script to
extract the model artifacts from their gzipped files
in the `models` test directory.

The non-gzip artifacts are ignored in the .gitignore,
as well as removed in the clean script.

* Copy WASM build artifacts to test directory in test-wasm.py

This commit taks the WASM artifacts generated by the
build script and copies them to a directory for use
in tests.

* Produce hash of generated JS in test-wasm.py

This commit computes a hash of the generated JavaScript,
since the test runner adds it to the worker global scope
using `eval`. This ensures that our test runner will only 
`eval` the intended script.

* Add Web Worker simulation infrastructure

This commit adds a minimal API surface of the WorkerGlobalScope
API functionality that we use for Translations within Firefox,
wrapping the Node.js worker_threads equivalent behavior underneath.

This allows us to test the generated code in a Node.js environment
with the same API calls that we use in Firefox.

* Add Translations Engine and worker implementation

This commit adds a simplified and minimal implementation of our
Translations Engine from the mozilla-unified source tree, which
is capable of starting a web-worker translator between a given
language pair and translating a single message at a time.

* Add test cases for current translations WASM bindings

Adds test cases that test the current translation functionality
end-to-end, including plaint-text translations and HTML translations.